fix: divide file
