+++ src/sections/ohisamaPage/index.ts
@@ -1,0 +1,3 @@
+export { OhisamaProfile } from './OhisamaProfile'
+export { LiveLog } from './LiveLog'
+export { QAndA } from './QAndA'
